Fix int32 pointer-arithmetic overflow in layer_norm/layernorm_gated/k_activations kernels
Triton's JIT specializes each operand of a pointer-arithmetic expression
(program-id, strides, etc.) as a 32-bit int based on that operand's own
value at specialization time, not the value their product will reach.
Standard integer promotion means the whole chained multiplication stays
32-bit as long as the leading operand is 32-bit, even if a later operand
is large -- once the true offset exceeds 2**31 - 1, it wraps to a
garbage address, either crashing with an illegal memory access or
silently reading/writing the wrong location.

These kernels compute row * stride_x_row where row = tl.program_id(0)
ranges over all flattened (batch*seqlen) rows, so an ordinarily large
row count is enough on its own to trigger the overflow, no
non-contiguous view needed. Cast the relevant program-id-derived index
to tl.int64 immediately after tl.program_id() in layer_norm.py,
layernorm_gated.py, and k_activations.py. Cheap, unconditional, no
measurable regression on the common contiguous/short-sequence case.

See https://github.com/state-spaces/mamba/issues/1015 for the full
writeup covering this same overflow class across the codebase,
including a concrete repro on a real production model.

Regression tests are guarded with @gpu_memory_skipif and checked with
the chunked assert_isfinite helper (both from overflow_test_utils.py),
since these allocate tens of GiB at the row counts needed to actually
trigger the overflow. Every test was verified to actually catch the
regression it targets, by temporarily reverting the relevant int64 cast
and confirming the test fails before restoring the fix.

+++ mamba_ssm/ops/triton/k_activations.py
@@ -31,7 +31,8 @@ def _swiglu_fwd_kernel(
     BLOCK_N: tl.constexpr,
 ):
     # Map the program id to the row of X and Y it should compute.
-    row = tl.program_id(0)
+    # int64 to avoid int32 overflow, see https://github.com/state-spaces/mamba/issues/1015
+    row = tl.program_id(0).to(tl.int64)
     start_col = tl.program_id(1) * BLOCK_N
     X += row * stride_x_row
     Y += row * stride_y_row
@@ -93,7 +94,8 @@ def _swiglu_bwd_kernel(
     RECOMPUTE_OUTPUT: tl.constexpr,
 ):
     # Map the program id to the row of X and Y it should compute.
-    row = tl.program_id(0)
+    # int64 to avoid int32 overflow, see https://github.com/state-spaces/mamba/issues/1015
+    row = tl.program_id(0).to(tl.int64)
     start_col = tl.program_id(1) * BLOCK_N
     X += row * stride_x_row
     Y += row * stride_y_row
